# apps/backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/v1/users": {
      "post": {
        "description": "Create user after Clerk signup, sync to DB",
        "summary": "Create a new user",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "createUser",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "user",
                      "vendor",
                      "delivery_partner",
                      "admin"
                    ]
                  },
                  "profilePicture": {
                    "type": "string",
                    "format": "uri"
                  }
                },
                "required": [
                  "email",
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    },
                    "phoneNumber": {
                      "type": "string",
                      "nullable": true
                    },
                    "role": {
                      "type": "string"
                    },
                    "isVerified": {
                      "type": "boolean"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "loyaltyPoints": {
                      "type": "number"
                    },
                    "profilePicture": {
                      "type": "string",
                      "nullable": true
                    },
                    "twoFactorEnabled": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "email",
                    "username",
                    "password",
                    "phoneNumber",
                    "role",
                    "isVerified",
                    "isActive",
                    "loyaltyPoints",
                    "profilePicture",
                    "twoFactorEnabled"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users/me/addresses": {
      "post": {
        "summary": "Add address",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "createAddress",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "label": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "detailAddressDirection": {
                    "type": "string"
                  },
                  "isDefault": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "label",
                  "latitude",
                  "longitude",
                  "phoneNumber",
                  "firstName",
                  "lastName",
                  "detailAddressDirection",
                  "isDefault"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "latitude": {
                      "type": "number"
                    },
                    "longitude": {
                      "type": "number"
                    },
                    "phoneNumber": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "detailAddressDirection": {
                      "type": "string"
                    },
                    "isDefault": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "userId",
                    "label",
                    "latitude",
                    "longitude",
                    "phoneNumber",
                    "firstName",
                    "lastName",
                    "detailAddressDirection",
                    "isDefault"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "List my addresses",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "listAddresses",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      },
                      "userId": {
                        "type": "string"
                      },
                      "label": {
                        "type": "string"
                      },
                      "latitude": {
                        "type": "number"
                      },
                      "longitude": {
                        "type": "number"
                      },
                      "phoneNumber": {
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "detailAddressDirection": {
                        "type": "string"
                      },
                      "isDefault": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "createdAt",
                      "updatedAt",
                      "userId",
                      "label",
                      "latitude",
                      "longitude",
                      "phoneNumber",
                      "firstName",
                      "lastName",
                      "detailAddressDirection",
                      "isDefault"
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/users/me/addresses/{id}": {
      "patch": {
        "summary": "Update address",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "updateAddress",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 20
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "isDefault": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "latitude": {
                      "type": "number"
                    },
                    "longitude": {
                      "type": "number"
                    },
                    "phoneNumber": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "detailAddressDirection": {
                      "type": "string"
                    },
                    "isDefault": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "userId",
                    "label",
                    "latitude",
                    "longitude",
                    "phoneNumber",
                    "firstName",
                    "lastName",
                    "detailAddressDirection",
                    "isDefault"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete address",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "deleteAddress",
        "responses": {
          "204": {
            "description": "204",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/users/me/addresses/{id}/default": {
      "patch": {
        "summary": "Set default address",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "setDefaultAddress",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "null",
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "latitude": {
                      "type": "number"
                    },
                    "longitude": {
                      "type": "number"
                    },
                    "phoneNumber": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "detailAddressDirection": {
                      "type": "string"
                    },
                    "isDefault": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "userId",
                    "label",
                    "latitude",
                    "longitude",
                    "phoneNumber",
                    "firstName",
                    "lastName",
                    "detailAddressDirection",
                    "isDefault"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/users/me/loyalty/balance": {
      "get": {
        "summary": "Get current balance",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "getCurrentBalance",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "balance": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "balance"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/loyalty/redeem": {
      "post": {
        "summary": "Redeem points",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "redeemPoints",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "points": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": 0
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 255
                  }
                },
                "required": [
                  "points",
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/loyalty/adjust": {
      "post": {
        "summary": "Adjust points (admin/system)",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "adjustPoints",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "pointsChange": {
                    "type": "number"
                  },
                  "transactionType": {
                    "type": "string",
                    "enum": [
                      "EARN",
                      "REDEEM",
                      "ADJUST"
                    ]
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 255
                  },
                  "performedBy": {
                    "type": "string"
                  },
                  "referenceId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "referenceType": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "userId",
                  "pointsChange",
                  "transactionType",
                  "reason",
                  "performedBy"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "Boilerplate REST API - Documentation",
    "description": "Boilerplate REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}